REF: Move relevant exceptions
skipci
